+++ pkg/cotd/entity.go
