+++ pkg/cotd/interface.go
