+++ pkg/meta/meta.go
@@ -1,7 +1,7 @@
 package meta
 
 type Meta struct {
-	Status string `json:status`
+	Status string `json:"status"`
 }
 
 func PublishMetaChanges(status Meta, sc chan<- Meta) {
