+++ pkg/mtg/entity.go
