+++ pkg/mtg/interface.go
