+++ pkg/mtg/testdata/b.png
