+++ pkg/mtg/testdata/broken_card.json
