+++ pkg/mtg/testdata/gideon_blackblade.json
