+++ pkg/mtg/testdata/ironclad_krovod.json
