+++ pkg/rakugaki/entity.go
