+++ pkg/rakugaki/interface.go
