+++ pkg/sakuga/entity.go
