+++ pkg/sakuga/interface.go
